+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version 0.1.1
 
 * Added functions latticeSize, findLatticeTimeIndex.
 
+* Added function estimateAt.
+
 
 Version 0.1
 -----